[ADD]변수명 변경 및 수정
