+++ src/models/product.dao.js
@@ -1,6 +1,6 @@
 const { appDataSource} = require('./data-source');
 
-const searchProduct = async(name) => {
+const getProductsByCategoryId = async(categoryId) => {
     return await appDataSource.query(
         `
     SELECT
@@ -14,7 +14,7 @@ const searchProduct = async(name) => {
     JOIN product_options po ON po.product_id = p.id
     JOIN product_option_images poi ON poi.product_option_id = po.id
     WHERE p.category_id = ?;
-    `, [name]);
+    `, [categoryId]);
 };
     
-module.exports = {searchProduct};
+module.exports = {getProductsByCategoryId};
